Update to allow use with `tox-uv`
Some issues were raised when putting the repo through a trial
run of `tox` with `tox-uv` installed.

"pip" is not included in package dependencies, but the pip wrapper,
`pip-install-dependency-groups` uses it. `tox` environments were still
using the wrapper.

To resolve, `tox.ini` is updated to use the new `dependency_groups`
support from v4.22.0+