Remove the 'dependabot canary' workflow
The idea here was to leverage dependabot to force an update, but it
doesn't appear to be working.

Dependabot isn't triggering in a reliable and timely manner, which
maybe could be fixed. But it's not clear to me that this is solving a
real problem -- the version bump script should be updated to fix this,
but even if it isn't, we can manually bump or rely on pre-commit.ci .

The thing which pushes me to remove it is that this interferes with
pre-commit.ci update PRs. (see #18)